fix: authentication issue on invitations page
